requirements.txt
scrapy==2.11.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pymongo==4.6.0
redis==5.0.1
celery[redis]==5.3.4
flower==2.0.1
textblob==0.17.1
vaderSentiment==3.3.2
schedule==1.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.14.2
streamlit==1.28.1
wordcloud==1.9.3
